Move diff view checkbox to column dropdown
