+++ mlflow/server/js/src/experiment-tracking/components/RunsTableColumnSelectionDropdown.js
@@ -1,7 +1,10 @@
 import React from 'react';
 import { Dropdown, Icon, Menu } from 'antd';
 import { Button } from '../../shared/building_blocks/Button';
+import { Checkbox } from '../../shared/building_blocks/Checkbox';
 import { SearchTree } from '../../common/components/SearchTree';
+import Utils from '../../common/utils/Utils';
+import ExperimentViewUtil from './ExperimentViewUtil';
 import PropTypes from 'prop-types';
 import _ from 'lodash';
 import { COLUMN_TYPES, ATTRIBUTE_COLUMN_LABELS } from '../constants';
@@ -13,6 +16,10 @@ export class RunsTableColumnSelectionDropdown extends React.Component {
     paramKeyList: PropTypes.array,
     metricKeyList: PropTypes.array,
     visibleTagKeyList: PropTypes.array,
+    runInfos: PropTypes.array,
+    paramsList: PropTypes.array,
+    metricsList: PropTypes.array,
+    tagsList: PropTypes.array,
     onCheck: PropTypes.func,
     categorizedUncheckedKeys: PropTypes.object.isRequired,
   };
@@ -25,6 +32,7 @@ export class RunsTableColumnSelectionDropdown extends React.Component {
 
   state = {
     menuVisible: false,
+    diffViewSelected: false,
   };
 
   handleCheck = (checkedKeys, allKeys) => {
@@ -100,6 +108,92 @@ export class RunsTableColumnSelectionDropdown extends React.Component {
     ];
   }
 
+  getCategorizedColumnsDiffView() {
+    const {
+      paramKeyList,
+      metricKeyList,
+      visibleTagKeyList,
+      runInfos,
+      paramsList,
+      metricsList,
+      tagsList,
+    } = this.props;
+    const attributeKeyList = [
+      ATTRIBUTE_COLUMN_LABELS.RUN_NAME,
+      ATTRIBUTE_COLUMN_LABELS.USER,
+      ATTRIBUTE_COLUMN_LABELS.SOURCE,
+      ATTRIBUTE_COLUMN_LABELS.VERSION,
+    ];
+    let attributes = [];
+    let params = [];
+    let metrics = [];
+    let tags = [];
+
+    for (let index = 0, n = runInfos.length; index < n; ++index) {
+      const paramsMap = ExperimentViewUtil.toParamsMap(paramsList[index]);
+      const metricsMap = ExperimentViewUtil.toMetricsMap(metricsList[index]);
+      const tagsMap = tagsList[index];
+
+      attributes.push([
+        Utils.getRunName(tagsList[index]),
+        Utils.getUser(runInfos[index], tagsList[index]),
+        Utils.formatSource(tagsList[index]),
+        Utils.getSourceVersion(tagsList[index]),
+      ]);
+      params.push(
+        paramKeyList.map((paramKey) => {
+          return paramsMap[paramKey] ? paramsMap[paramKey].getValue() : '';
+        }),
+      );
+      metrics.push(
+        metricKeyList.map((metricKey) => {
+          return metricsMap[metricKey] ? metricsMap[metricKey].getValue() : '';
+        }),
+      );
+      tags.push(
+        visibleTagKeyList.map((tagKey) => {
+          return tagsMap[tagKey] ? tagsMap[tagKey].getValue() : '';
+        }),
+      );
+    }
+    // Transpose the matrices so that we can evaluate the values 'column-based'
+    attributes = _.unzip(attributes);
+    params = _.unzip(params);
+    metrics = _.unzip(metrics);
+    tags = _.unzip(tags);
+    const allEqual = (arr) => arr.every((val) => val === arr[0]);
+
+    return {
+      [COLUMN_TYPES.ATTRIBUTES]: attributeKeyList.filter((v, index) => {
+        return allEqual(attributes[index]);
+      }),
+      [COLUMN_TYPES.PARAMS]: paramKeyList.filter((v, index) => {
+        return allEqual(params[index]);
+      }),
+      [COLUMN_TYPES.METRICS]: metricKeyList.filter((v, index) => {
+        return allEqual(metrics[index]);
+      }),
+      [COLUMN_TYPES.TAGS]: visibleTagKeyList.filter((v, index) => {
+        return allEqual(tags[index]);
+      }),
+    };
+  }
+
+  handleDiffViewCheckboxChange = () => {
+    const { onCheck } = this.props;
+    this.setState({ diffViewSelected: !this.state.diffViewSelected }, () => {
+      const categorizedUncheckedKeys = this.state.diffViewSelected
+        ? this.getCategorizedColumnsDiffView()
+        : {
+            [COLUMN_TYPES.ATTRIBUTES]: [],
+            [COLUMN_TYPES.PARAMS]: [],
+            [COLUMN_TYPES.METRICS]: [],
+            [COLUMN_TYPES.TAGS]: [],
+          };
+      onCheck(categorizedUncheckedKeys);
+    });
+  };
+
   handleSearchInputEscapeKeyPress = () => {
     this.setState({ menuVisible: false });
   };
@@ -109,9 +203,14 @@ export class RunsTableColumnSelectionDropdown extends React.Component {
   };
 
   render() {
-    const { menuVisible } = this.state;
+    const { menuVisible, diffViewSelected } = this.state;
     const content = (
       <Menu style={{ maxHeight: 480, overflowY: 'scroll' }}>
+        <Checkbox
+          label='Diff View'
+          isSelected={diffViewSelected}
+          onCheckboxChange={this.handleDiffViewCheckboxChange}
+        />
         <SearchTree
           data={this.getData()}
           onCheck={this.handleCheck}
